Stage 2
I have probably spend about 6 hours total now on this program. 

In this version, the vehicles have been given DNA. They are now spawned either more attracted to the food, or the poison. The ones that are better at detecting food with a small perception of poison should be the vehicles that survive the longest. 

Also, now they are repelled by the edge of the screen, and should always be focused around the centre. 

In the next version I am looking to add the actual evolution. The vehicles will be able to pass down their DNA onto other vehicles that will receive a mutated version of that DNA which hopefully makes them slightly better than their parent. 

Exciting stuff, will see how it turns out.