Stage 3
In this stage, I have added the ability for the vehicles to reproduce over time, they have a set chance of producing offspring if they survive for a determined amount of time. Also, to increase longevity, wherever a vehicle dies, new food will appear.

In the next version of this project, I will be adding a feature that allows the DNA passed from parent to child to mutate, thus creating a more realistic representation of evolution.