Stage 5
This stage concludes the functionality of this project. The vehicles are able to hunt down the food and avoid the poison (if given the correct DNA), the ones that survive the longest will then pass down their DNA. The DNA passed down will have a mutation factor, so hopefully their offspring can be even better suited to avoiding the poison and eating the food.

In the next stage I will be adding a predator class. The predators will need to hunt the vehicles to survive.

Hopefully this all turns out well!